+++ src/components/OverrideModal/OverrideModal.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
+import { omit } from 'lodash';
 
 import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import {
@@ -18,14 +19,16 @@ import { DueDatePicker } from '@folio/stripes/smart-components';
 import {
   DATE_PICKER_DEFAULTS,
   INVALID_DATE_MESSAGE,
+  ITEM_NOT_LOANABLE,
+  MAX_ITEM_BLOCK_LIMIT,
 } from '../../constants';
 
 function OverrideModal(props) {
   const {
     stripes,
-    overrideModalOpen,
     closeOverrideModal,
     onOverride,
+    message,
     item: {
       title,
       barcode,
@@ -36,20 +39,42 @@ function OverrideModal(props) {
   const [comment, setAdditionalInfo] = useState('');
   const [dueDate, setDatetime] = useState('');
 
+  const itemIsNotLoanable = message.includes(ITEM_NOT_LOANABLE);
+  const blockLimitIsReached = message.includes(MAX_ITEM_BLOCK_LIMIT);
+
   const handleDateTimeChanged = (newDateTime) => {
     setDatetime(newDateTime);
   };
 
-  const canBeSubmitted = comment && dueDate !== INVALID_DATE_MESSAGE;
+  const canBeSubmitted = itemIsNotLoanable
+    ? comment && dueDate !== INVALID_DATE_MESSAGE
+    : comment;
+
+  const getModalLabel = () => {
+    let label = '';
+    if (itemIsNotLoanable) {
+      label = <FormattedMessage id="ui-checkout.overrideLoanPolicy" />;
+    }
+
+    if (blockLimitIsReached) {
+      label = <FormattedMessage id="ui-checkout.overrideItemBlock" />;
+    }
+
+    return label;
+  };
 
   const onSubmit = async (event) => {
     event.preventDefault();
     closeOverrideModal();
-    onOverride({
+    const overrideItem = {
       comment,
       dueDate,
       barcode,
-    });
+    };
+
+    return itemIsNotLoanable
+      ? onOverride(overrideItem)
+      : onOverride(omit(overrideItem, 'dueDate'));
   };
 
   return (
@@ -58,8 +83,8 @@ function OverrideModal(props) {
       dismissible
       enforceFocus={false}
       data-test-override-modal
-      open={overrideModalOpen}
-      label={<FormattedMessage id="ui-checkout.overrideLoanPolicy" />}
+      open
+      label={getModalLabel()}
       onClose={closeOverrideModal}
     >
       <form
@@ -74,6 +99,7 @@ function OverrideModal(props) {
             />
           </p>
         </Col>
+        {itemIsNotLoanable &&
         <Col
           xs={12}
           data-test-override-modal-due-date-picker
@@ -82,19 +108,24 @@ function OverrideModal(props) {
             required
             initialValues={DATE_PICKER_DEFAULTS}
             stripes={stripes}
-            dateProps={{ label: (
-              <FormattedMessage id="ui-checkout.cddd.date">
-                {message => `${message} *`}
-              </FormattedMessage>
-            ) }}
-            timeProps={{ label:(
-              <FormattedMessage id="ui-checkout.cddd.time">
-                {message => `${message} *`}
-              </FormattedMessage>
-            ) }}
+            dateProps={{
+              label: (
+                <FormattedMessage id="ui-checkout.cddd.date">
+                  {label => `${label} *`}
+                </FormattedMessage>
+              )
+            }}
+            timeProps={{
+              label: (
+                <FormattedMessage id="ui-checkout.cddd.time">
+                  {label => `${label} *`}
+                </FormattedMessage>
+              )
+            }}
             onChange={handleDateTimeChanged}
           />
         </Col>
+        }
         <Col
           data-test-override-modal-comment
           xs={12}
@@ -106,6 +137,14 @@ function OverrideModal(props) {
           />
         </Col>
         <Col xs={12}>
+          {itemIsNotLoanable &&
+            <>
+              <br />
+              <br />
+              <br />
+              <br />
+            </>
+          }
           <Row end="xs">
             <Button
               onClick={closeOverrideModal}
@@ -124,10 +163,6 @@ function OverrideModal(props) {
           </Row>
         </Col>
       </form>
-      <br />
-      <br />
-      <br />
-      <br />
     </Modal>
   );
 }
@@ -135,9 +170,9 @@ function OverrideModal(props) {
 OverrideModal.propTypes = {
   stripes: stripesShape.isRequired,
   item: PropTypes.object.isRequired,
-  overrideModalOpen: PropTypes.bool.isRequired,
   onOverride: PropTypes.func.isRequired,
   closeOverrideModal: PropTypes.func.isRequired,
+  message: PropTypes.string.isRequired,
 };
 
 export default OverrideModal;