+++ src/constants.js
@@ -30,7 +30,12 @@ export const defaultPatronIdentifier = 'BARCODE';
 
 export const ITEM_NOT_LOANABLE = 'Item is not loanable';
 
-export const OVERRIDABLE_ERROR_MESSAGES = [ITEM_NOT_LOANABLE];
+export const MAX_ITEM_BLOCK_LIMIT = 'Patron has reached maximum limit of';
+
+export const OVERRIDABLE_ERROR_MESSAGES = [
+  ITEM_NOT_LOANABLE,
+  MAX_ITEM_BLOCK_LIMIT,
+];
 
 export const statuses = {
   CHECK_OUT: 'Check out',