+++ test/bigtest/constants.js
@@ -1,5 +1,7 @@
 export const loanPolicyId = 'd9cd0bed-1b49-4b5e-a7bd-064b8d177231';
 export const loanPolicyName = 'testLoanPolicyName';
+export const loanPolicyWithLimitId = 'd9cd0bed-1b49-4b5e-a7bd-064b8d177232';
+export const loanPolicyWithLimitName = 'loanPolicyWithLimitName';
 export const notLoanablePolicyName = 'Not loanable';
 export const notLoanablePolicyId = 'notLoanablePolicyId';
 export const notLoanableItemBarcode = 'notLoanableBarcode';
@@ -8,3 +10,6 @@ export const blockedUserId1 = 'blockedUserId1';
 export const blockedUserId2 = 'blockedUserId2';
 export const blockedMessage = 'Patron has reached maximum allowed number of items charged out';
 export const manualBlockMessage = 'Invalid email and mailing addresses';
+export const barcode = '555555';
+export const userBarcode = '123456';
+export const barcodeWithLimitLoanPolicy = 'barcodeWithLimitLoanPolicy';